+++ src/pages/login.tsx
@@ -1,27 +1,14 @@
 import React, { FC } from 'react'
-import { Box, Typography } from '@material-ui/core/'
-import { makeStyles, createStyles } from '@material-ui/core/styles'
+import { Box } from '@material-ui/core/'
 import Emoji from '../components/Emoji'
 import TextField from '../components/TextField'
 import PasswordTextField from '../components/PasswordTextField'
 import Layout from '../components/Layout'
 import Titulo from '../components/Titulo'
 import InputLabel from '../components/InputLabel'
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    contato: {
-      marginTop: 16,
-      width: 320,
-      fontSize: 18,
-      lineHeight: '25px'
-    }
-  })
-)
+import EsqueciMInhaSenha from '../components/login/EsqueciMinhaSenha'
 
 const Login: FC = () => {
-  const classes = useStyles()
-
   return (
     <Layout textoBotao="Entrar">
       <Titulo>
@@ -29,16 +16,15 @@ const Login: FC = () => {
       </Titulo>
 
       <Box mt={5}>
-        <InputLabel>Email</InputLabel>
-        <TextField />
+        <Box>
+          <InputLabel>Email</InputLabel>
+          <TextField />
 
-        <InputLabel>Senha</InputLabel>
-        <PasswordTextField />
+          <InputLabel>Senha</InputLabel>
+          <PasswordTextField />
+        </Box>
 
-        <Typography className={classes.contato}>
-          Esqueceu sua senha? Entre em contato com a gente pelo email
-          contato@jornadasolar.com.br
-        </Typography>
+        <EsqueciMInhaSenha />
       </Box>
     </Layout>
   )
